third_party/WebKit/LayoutTests/fast/dom/Node/append-child-error.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
</head>
<body>
This test case verifies if relevant TypeError is thrown when invalid Node argument is given.
<script>
test(function() {
    assert_throws(new TypeError(), function() { document.body.appendChild() })
    assert_throws(new TypeError(), function() { document.body.appendChild(null) })
    assert_throws(new TypeError(), function() { document.body.appendChild({'a':'b'}) })
}, "Node.appendChild should throw TypeError when a wrong type of argument is passed.")
</script>
</body>
</html>
